Back-End/requirements.txt
Django~=5.1.3
django-environ
channels~=4.1.0
channels_redis
psycopg2
dj_database_url
paho-mqtt~=2.1.0
djangorestframework~=3.15.2
python-dotenv~=1.0.1
pymongo~=4.10.1
django-heroku~=0.3.1
asgiref~=3.8.1
pytz~=2024.1
django-cors-headers
django-extensions
whitenoise
daphne==4.1.2
celery
djangorestframework-simplejwt~=5.3.1
boto3~=1.35.76
firebase-admin~=6.6.0